various: add resource-id-ing ~| into common paths
We see crash traces for all kinds of things. Currently, we print poke
and watch nack traces to the user's terminal. These by themselves only
point at the codepath, but don't explicate the crash reason (which could
also be deduced from the trace) or the resource that was operated on
(which might not be recoverable in cases where we don't have/print the
wire).

Here, we add a couple of trace hints, to print subscription paths and
resource identifiers in common crash pathways.