notify: add +notify-delivery-report for push stats
Adds a generator that displays some basic statistics about notifications
that exist in the notify agent's state, such as the amount that were
requested by the client, and the average amount of time it took for the
client to request them.

Note that this will be misleading when using clients that haven't yet
been updated to use the new endpoints, as without those, notifications
won't be marked as having been requested/served.

Example output:

Between ~2024.6.27 and ~2024.6.28,
Served 3 out of 4 notifications (75%).
Average time-to-request was ~s15.