activity: only direct kids from +get-children
Previously, we were giving all sources as children of base. But this
doesn't fit with our source hierarchy, and is not how +get-children is
used in /app/activity.

Now, we only return %group and %dm sources as children of %base, and let
the code iterate further on those for deeper processing.

With this change, we no longer do repeated processing of %thread and
%dm-thread sources, resulting in (on ~palfun) a 20% speedup in unread
summarization.