docket-proxy: mitm docket to inject custom headers
Some of the upcoming frontend code needs access to functionality the
browser only enables in the presence of specific headers on the
responses serving the source code files.

Docket serves our frontend files, but does not have any affordances for
specifying custom headers. But it serves from globs, a non-trivial task
to take over, even if we were to run a custom fork of it.

Instead, here we introduce a local proxy that watches on the endpoint
docket would normally serve on. It passes requests through unaltered,
but injects the desired headers into responses before sending them back
to the client.

This approach will stop working if docket ever injects responses on this
endpoint into the eyre cache, which means we can't trivially capture
them in userspace anymore. However, vere 3.1 never ever, so we should be
good for the time being.