channels: don't crash during "double join"
Previously, if we tried to join a channel we already had state for, we
would crash.

Here, we do a check on our subs, reestablish them if needed, but
otherwise no-op. This lets us import channel state from backups without
it interfering with subsequent group rejoins. It also speeds up the
rejoins: we simply start fetching updates again from where we the
imported state left off.