imports: ensure imports happen in explicit order
We apply the $pimp pattern to other agents, so that we can enforce
import order of activity -> channels-server -> channels -> groups.

We care about the order being that way, because channels might subscribe
to channels-server, and groups will tell channels where to resubscribe,
but those subscriptions should key off the data we already have instead
of starting from a clean slate.