channels-server: send wake alert after import
When importing from a backup, send all the subscribers we remember a
poke so that they can try re-establishing their subscriptions.

Doing this here also means that in the groups wake case, we can stop
doing this for all channels in the group, which might not be accurated
for mixed group-host channel-host scenarios.