expose: refresh entire cache on-load
And defer this to a separate event, because rendering scries, which is
unsafe during load.

We make sure to update the cache entries for all posts, because style
files might've changed that impact their pages.

...This may actually be somewhat sluggish in cases where you have a lot
of content exposed. We should consider just purging the cache for
individual pages instead, and only re-rendering and caching those
on-demand.