expose: don't render links inside widget previews
Apparently nesting <a> tags is illegal, and browsers will actively
malform your document tree to avoid having to render such cases.

To avoid that, we add an optional flag to the +en-manx:channel-utils
core that can be set to false to disable the rendering of anchor
elements. For images, they won't be wrapped inside an <a> tag at all.
For text links, we wrap them in a span.faux-a instead, so that they may
still be styled to look similar to real anchors.