emojimart: transform "shortcodes" back into emoji
Instead of the raw unicode, we've stored reacts as ascii-textual
representations, which are largely impractical, require a library for
rendering, and end up with rather ridiculous representations, such as
:hand_with_index_finger_and_thumb_crossed::skin-tone-2: in place of 🫰🏻.

Here, we port part of the aforementioned rendering library into hoon.
Specifically, the mappings of those "shortcode" identifiers to their
unicode characters, and a helper function for looking those up from a
potentially with-modifier shortcode.

Note the auraless atoms produced here, because there exists no aura for
"unicode character". (`@c` is a UTF-32 codepoint, not a semantic
character.)

At some point, we may want a more full-fledged unicode wrangling library
in hoon. But for now, this suffices.

Includes 80% of migration code, which was used for performance testing.
Building the library file takes half a minute, running the migration
takes a mere second.