contacts: memoize on smaller subject
We don't actually care about the whole bowl, we just take it for
ergonomics. The memoization shouldn't account for the parts of the bowl
that we don't use. This way, if different agents scry during the same
event (or the same agent on separate invocations) we still pull the
rolodex from cache.