expose: re-render cache if local profile changes
This is a detail that appears on most of the pages that we publish, that
the user has control over, so we should reflect is accurately in a
timely fashion.

We don't care too much right now about cache busting in response to
other people's profile detail changes, because we don't display
comments/replies, and publishing someone else's content isn't a
primary/supported use-case. If we ever do want that, probably better to
just periodically refresh the cache.