verifier: pin-based urbit registration
The previous intended design, of needing to register two-way, prevents
you from registering multiple urbit ids. This still only lets every
urbit id be claimed by one registrant, but that's consistent with all
other identifiers.

Arguably it would be better for the urbit being registered to sign a
claim instead, to the effect that the registrant controls it. Explicit
input of the registrant's urbit id also prevents pin confusion attacks,
where the pin is correct but the registrant isn't who you expect.

We provide this implementation as a proof of concept for now, and
continue on with more important things.