tests: ensure pre-initialization invocations crash
Previously, if you forgot to initialize the agent core to test against,
we would run with the bunt of $agent:gall, which is a core that just
silently succeeds.

Here, we put a skeleton agent into the continuation state, so that any
invocations crash (and show a high-priority printf warning you of this
fact).