verifier: phone number support
Talks to a Tlon Hosting API for doing verification based on a OTP text
message. Shortcuts this whole process if Hosting has previously verified
this phone number for the ship that's registering it.

Includes a new +merge helper in the testing code that proceeds (cached)
into a shared continuation for each branch of the test.