verifier: sign both "half" and "full" attestations
The existing attestation includes a signature over a "full" message,
containing all specifics of the identfier that was verified. But in some
cases, it may be desirable to share a verifier attestation without
revealing the full identifier. For example, wanting to substantiate the
claim that you have registered a phone number, without revealing the
number in question.

Here, we make the verifier sign a "half" message as well, which includes
the identifier kind, but no other details about it. This lets the client
choose between the two signatures as appropriate.