lanyard: inject verified ids into contacts profile
When we have verified identifiers, inject them into the local contacts
profile. The details of what exactly we put in differs per kind of
identifier.

For phone numbers, we only inject a single "have registered any phone
nr" attestation, along with its timestamp.

For urbits, we inject the set of urbits that we've registered, along
with the attestation for each.

Currently does this regardless of per-id discovery settings. Publishing
control here should probably go through different setings, but the
current behavior is fine for defaults.