verifier: make sure initial fact is complete
Because we only mark reverse-lookup ownership of completed
registrations, the initial subscription response was failing to include
in-progress registrations.

Here, we put in a quick fix, but note that it might be better to change
the semantics of the .owners lookup. Performance of the implementation
as-is will degrade for large numbers of registrations.