verifier, lanyard: share & store service urls
Verifiers now publish a base url at which their service is accessible
over http. This lets clients construct clearweb-accessible urls at which
attestations can be checked for legitimacy.

Also updates lanyard to store these base urls and include them whenever
we inject details into the contacts profile.

The situation around lanyard's subscriptions changes a little bit. It no
longer auto-initiates contact with the default service, instead always
establishes subscriptions on an as-needed basis when pokes are about to
be made to a verification provider. The .ledgers in state track all
known providers. An entry there indicates we intend and expect to have
subscriptions open.

(Note that this commit does not actually include the implementation of
the http serving that this helps hook up. More to follow.)