verifier: serve details on attestation pages
Accessible at /attestations/[signature], the service now provides pages
that show the details of the attestation corresponding to the signature.
The level of detail shown (specifically, whether the actual identifier
is shown) depends on whether the signature is the "half" or "full"
signature for that identifier.

This should not provide additional abuse surface area. Signatures are
only provided to the registrant, where/how to share those is up to them.
Signatures are 512 bits, not trivially crawled. We should implement
rate-limiting on the HTTP requests regardless.