lanyard: give fact about eager %wait after %start
We eagerly set the state in a record to %wait after we first %start it.
Here, we make sure we send that out as a fact, so any subscribers know
right away.

(Good chance the verifier will also set the status to %wait, resulting
in a second fact, but that should not be a problem.)