lanyard: store query results indefinitely
Instead of storing only the queries themselves in state, and deleting
them after we've emitted an update containing their result, we now track
both the query and the response in state indefinitely.
(These _should_ get cleaned up based on some heuristic, but we punt on
that for now.)

This will let clients obtain query results without strictly needing to
be listening at the time we get the result. (Though relevant scry
endpoints are yet to be added...)