lanyard: check signatures locally before querying
For %valid-jam queries, which contain the whole $urbit-signature, we
have enough information to try checking the signature locally first,
before we ask the signer (verifier) over the network.

If we know the public key corresponding to the given ship and life, we
get it out of jael and validate the signature with it.
If it's invalid, we don't need to ask the signer. If we don't have the
necessary pubkeys, we strictly need to ask the signer. If the signature
is valid, that's a good first indication to show the user, but we still
need to check with the signer in case it's been revoked.