profile: read contacts fields safely
types. It checked for the tags on values found in the map. However, it's
possible for values to be null, in which case trying to grab the head
would crash at runtime.

Here, we update the typecheck to check on the full value as opposed to
just the head.

Since this crashing code might have run in response to a contacts
subscription update, killing the subscription, we add a bit of logic to
+on-load to ensure that the contacts subscription gets set up again if
needed.