verifier: support bulk "whose" lookup requests
Implements a rate-limited batch lookup query in the style of
"Differential Contact Discovery" (from Hagen et al. (2022)).

The service stores a salted hash of the client's last-requested set. The
client reminds the service of the salt and set with each request, and
additionally provides additions and deletions on that set for
constructing the new request. Provided that the provided salt and set
match the hash stored in the service, the client only pays the
rate-limiting cost for the additions to the set. All other results are
"refreshed" for free.

This works well for legitimate cases, where the client asks about an
entire contact book: that set may contain many entries but changes
only a small amount (if at all) between requests. The abusive case, of
wanting to request as many unique entries as possible, will instead hit
rate-limits quite quickly.