verifier: twitter account verification support
To register a Twitter/X.com handle, the registrant is given a nonce.
They generate an $urbit-signature noun containing the handle and nonce,
jam it, post a tweet, and provide the verifier with the tweet ID. The
verifier then uses the x.com API to retrieve the tweet body, parses out
the jam, verifies its contents and signature, and approved the
registration as normal.