verifier: log usage & failures to posthog
Using the new /lib/logs. We make sure to log unexpected failures, and
key points in the registration flow: starting, completing, revoking, and
uncommon unhappy path cases like the user submitting an incorrect OTP
code.

We update the tests to ignore any and all logging cards, so that we can
just focus on the core behavior during the tests.