lanyard: contacts profile controls
Lets the user specify what verified identifier should be included in
their contacts profile. Defaults to exclude for each identifier, can be
set to include either the "half" (generic) or "full" (specific)
attestation.

Now with a non-temporary contact profile value key scheme, with the
%lanyard prefix followed by the id-kind, an index number for that kind,
and any of "since", "sign", "url" and "value" as applicable, all
separated by a hyphen.