groups: avoid %kick-ing empty list of paths
Kicking empty list of paths means "kick subscription that triggered this
event". This does mean that when constructing a path list of unknown
size, you have to catch the empty list case to avoid falling into that
specialized behavior.

Here, we check for the empty list, and avoid emiting kicks (or facts) in
that case.