groups: ignore %kicks for deleted groups
A %kick on a subscription to a group that had been deleted
would crash. This caused a problem in conjunction with the use
of lib-negotiate: during first loading the library sends a %kick on
every subscription in order to force protocol negotiation.
If a subscription was not properly cleaned-up by the agent, the
%kick will trigger a crash, interrupting the update.