de-html: forked +de-xml for improved tolerance
A modified version of the +de-xml:html parsers that has broader support
for real-world html, as opposed to the much more strict xml syntax.

This does not claim to be a fully-fledged html parser. We have presently
only used it for parsing the <head> sections of html documents. We do
not presently aim to support parsing arbitrary webpages reliably.