verifier: don't crash when rate-limiting
Instead, re-issue the current status with a 'rate limited' reason. This
lets us inform the client that what they did was correct but
soft-rejected (instead of hard-rejected like invalid pokes would be),
and also lets us log the rate limiting "event" so we can track the
frequency of their occurence (outside of the nack logging lanyard does).