verifier: log if 404 because of bad lookups
We already log when serving 404 responses, but those could occur due to
legitimate usage. However, if we serve a 404 because we looked up a
pointer that doesn't match the thing it's pointing to, that's bad
bookkeeping, indicative of a bug in our code, and we should log it as
such.