verifier: don't persist log details across events
Because the .log is in the agent core's context, it would get carried
along across invocations, as would any changes we made to that core.

Because we set some of those door arguments for logging purposes and
never _unset_ them, logs would _always_ have the optional values, just
set to the most recent value, unless they changed them themselves.

This is the smallest viable fix. Arguably this is paying for cuteness by
being even cuter. The logging pattern here remains really nice though,
so it's a bit of a wash...