expose: inject references into contacts profile
Adds exposed references into our contacts profile with the prefix
'expose-cite-' followed by a 0-based index number. Those fields will
always contain the reference path as a string.

This way, "pins" are shared over the network and can be displayed on
in-app profiles, assuming the viewer can resolve the references.