channels: add "targeted" reference fetching path
Adds a subscription path for resolving a reference through the specified
target ship (as opposed to always asking the channel host). This lets
one ask multiple sources in parallel, for improved resolution
performance and reliability. (ie, you might commonly want to ask both
the host, and the person who posted the reference.)

The new endpoint (/v3/said/etc) maintains the old behavior of giving
data from local if we know it, and kicking after a response has been
given.