expose: prime cache for pins when we see them
Expose was already subscribed to the contacts agent. Here, we update
that to use the latest subscription endpoint, and when hearing about
contacts whose profile contains pinned messages, we make a request to
the channels agent to fetch those messages, to prime the local cache, by
asking the ship that pinned them.