channels: retain mark types across protocol change
The recursive changes to the $inline type (specifically the %sect
addition) and the presence of both $inline and $block in $listing had
gone unnoticed, resulting in mark types changing unintentionally.

Here, we make sure all affected types get a proper entry in +v7 in
/sur/channels, and that mark files use that older type version as
necessary.

Mandatory utils and json changes included.

Note that the %channel-update and %channel-checkpoint marks, which are
emitted by channels-server, still change due to the protocol changes.
Since these marks only ever come out of channels-server, and interaction
with that agent is (intended to be) gated behind version negotiation, we
just let that happen.