metagrab: detect & indicate when we get blocked
Overly aggressive bot/scraper detection might serve 401s or 403s or 429s
or similar when they detect a request is being made in an unusual way.
Here, we add in the first couple of heuristics for detecting that case,
so that we may leave a little indicator for the client, who might be
able to handle those cases with fancier instruments.