lanyard: send invite links to clearweb verifiers
Whenever we complete a verification, or whenever a service changes its
clearweb base url, we send the service our personal invite link. The
latter also triggers during the service's +on-load, meaning we'll
periodically refresh the urls in case it changed, or failed last time.
This is preferable to elaborate retry and success tracking mechanism,
which add complexity and are prone to failure. (Timers in particular may
get lost in userspace.)