metagrab: fetch HEAD before doing a full GET
We only care about the response body when it's an html page. In all
other cases, we discard the body entirely and only extract data from the
headers.

Some urls might point to very large resources. Past a certain point, the
response is too big for vere to handle properly, and it will die.

Vere-side fix for that is forthcoming, but it's also not necessary for
metagrab to request the full responses in the first place. Instead, a
HEAD request will suffice.

Here, instead of going straight to a GET request, we always make a HEAD
request first. If the response has a text/html content-type, we proceed
with the GET request as normal. Otherwise, we make do with the headers.