chat: include sequence numbers
Adds sequence numbers to chat $writs. Includes all the relevant api
bumps. Note that this only affects client-facing apis. Inter-ship
communication is unaffected and does not get bumped.

In practice, from a json perspective, the new types are backwards
compatible: the only change is an additional 'seq' field inside the
$seal object.