chat: don't crash if writs page is empty
Previously, we would forcibly take the head and rear of the list of
writs. In cases where we produce the empty list of writs, this would
crash.

Now, in those cases, we leave out the "newer" and "older" pointers,
since we don't have any messages to use as a reference point.