Prevent case where boundary callbacks get blocked due to old metrics
Not 100% sure of this, but it looks like in some cases (especially when
loading lots of content quickly, causing browser slowdown), the
`isElementScrolledNearBottom/Top` functions return the wrong value,
likely due to `scrollHeight`/`scrollTop` that don't reflect the actual
UI state. This would prevent `onStartReached` from being called until
the content changes (e.g. loading more posts in the other direction),
which was a very bad experience.

This change forcibly resets the `useDeduplicateInvocationBy` key when
scrolling away from the scroll boundary: this way, scrolling away from
the boundary and then back to it will always call boundary callback
(e.g. `onStartReached`).

This does mean that a slow page load could lead to the user queueing up
multiple `onStartReached` as they impatiently scroll up and down - I
think it's reasonable to handle this in the layer above this view.