reel: +on-watch %fact only to current subscriber
Instead of sending the fact to _all_ subscribers, the fact should only
go to the subscription that was just now opened. Existing subscribers
should've already received this fact, and don't need to receive it again
until its content changes.

Because tlon app (ab)uses this subscription during reconnect logic, and
because of an eyre bug that fails to clean up subscriptions for expired
channels, some ships have many of these subscriptions open. Sending a
fact to all subscribers would cause a bunch of essentially no-op fact
processing logic to kick off.

Fixes for those issues incoming, but this one should be good to have
regardless.