channels/server: negotiate correct sequence nrs
During the migration, channels agent just assumed it had full backlog,
and counted up from 1. However, it might have had partial backlog, in
which case its count and numbering will be incorrect.

Here the channels agent triggers pokes for every channel it's aware of,
asking the host to send the correct set of sequence numbers over the
wire in a batch. We stagger the pokes pseudorandomly to avoid putting
undue pressure on the host, though in practice we don't expect this to
matter much.

If the host hasn't updated yet they will nack the poke, in which case we
set another timer to try again later. We keep trying indefinitely until
we get a response, to guarantee that we'll have eventually-correct
sequence numbers.

This "side-channel" approach (as opposed to baking it into the formal
agent protocol) lets us push this out as a hotfix as opposed to a
protocol upgrade.