chat: call explicitly-versioned types & json utils
This will save us from having to update the mark in the future. The mark
will not change, but we might add new versions. Accessing them from
"latest" would necessitate changing the mark code.

To facilitate this, we add /sur/chat-5 as a simple alias to be replaced
later. This should have lower impact and chance for mismanagement than
moving /sur/chat in its entirety or adding a symlink.