channels: retrieve tombstones explicitly
Instead of having the server send them over the subscription pathway,
have subscribers recover tombstones explicitly as part of the sequence
number "negotiation" flow.

Subscriptions aren't strictly guaranteed to be "live", so subscribers
could fall through the cracks. This might be superfluous work for
subscribers that hit the patched upgrade path, but really it's not much
work at all, we need to fetch sequence numbers anyway.