web: use web-native list for chat channels (#4868)
* Put FlatList element behind app-specific PostList component

* [wip] Add web-native PostList implementation

* Fix PostList content not filling container width

* Implement scroll anchoring for browsers without overflow-anchor

* Change "at bottom" -> "near bottom"

* Match existing onScrolledToBottom margin behavior

* Force scroll boundary thresholds explicit

* Organize code / add comments

* Use flex=1 instead of width=100% for web compatibility

it seemed like width="100%" was using a different reference on web than
on native, causing the author row to push out past its parent's
intrinsic max size. flex=1 seems to be the desired behavior anyways.

* Add switch to fixture to automatically load on scroll boundaries

* Fix stuck pagination on tall viewports

* Move relevant code into hook

* Pull hacky getter into function

* Remove unneeded(?) scroll-to-anchor reset on scroll-to-bottom

* Fix type error

* Stop setting initialNumToRender on PostList FlatList

* Fix lint

* Put web-specific scroller behind feature flag

* Pass `mode` prop to switch between trigger/auto-popup message actions

We were using `useIsWindowNarrow` to switch here, which didn't work for
narrow desktop layouts. In practice, we mount this component in
different places when handling hover-capable clients vs devices that
open context menus via long-press; so a static prop works fine for now.

* Prevent post actions trigger adding to ancestor scroll height

* Fix short chats not aligning to bottom

* Fix type

* Only run JS callback when needed

* Add `useFixtureFeatureFlag` helper to toggle flags in fixture

* Scroll using `scrollBy` instead of calculating absolute scroll

Calculating absolute scroll offset using `scrollTop + calculatedOffset`
caused visual jumps when scrolling using a trackpad, since the scroll
was being performed after the measured `scrollTop` became stale. Passing
the calculated offset directly to the `scrollBy` method (which scrolls
by a delta instead of an absolute position) fixes this jumpiness.

* Memoize callback

This was causing `onInitialScrollCompleted` to be called multiple times
on PostListFlatList

* Wrap callback in useCallback to avoid infinite render loop

* Stop re-rendering on IntersectionObserver's intersectionSet change

* Add debug flag to FORCE_MANUAL_SCROLL_ANCHORING

* Prevent case where boundary callbacks get blocked due to old metrics

Not 100% sure of this, but it looks like in some cases (especially when
loading lots of content quickly, causing browser slowdown), the
`isElementScrolledNearBottom/Top` functions return the wrong value,
likely due to `scrollHeight`/`scrollTop` that don't reflect the actual
UI state. This would prevent `onStartReached` from being called until
the content changes (e.g. loading more posts in the other direction),
which was a very bad experience.

This change forcibly resets the `useDeduplicateInvocationBy` key when
scrolling away from the scroll boundary: this way, scrolling away from
the boundary and then back to it will always call boundary callback
(e.g. `onStartReached`).

This does mean that a slow page load could lead to the user queueing up
multiple `onStartReached` as they impatiently scroll up and down - I
think it's reasonable to handle this in the layer above this view.

* stash

* [wip] perform scroll updates outside of React effects

* Handle list content changing size asynchronously without overflow-anchor

* Remove unused code

* Clean up ts-ignore

* Use scroll height instead of content to debounce boundary callbacks