groups: state migration part B - migrate requests in flight
After migrating the agent state, we need to migrate requests
that were in flight during the upgrade, but whose sequences are
altogether changed. For each foreign group (formerly gang), we re-issue
the request whenever possible, otherwise we signal an error.

For each hosted group, we send new invitations to each ship in the pending list.