chat: stop sending %epic facts to all subs
/lib/saga's +load trawls incoming subscriptions and gives an %epic fact
to every subscription coming in from outside. Aside from being a legacy
codepath, this rubs up against the /lib/discipline integration, which
wants every subscription path to specify its potential marks.

Instead of adding %epic to every subscription endpoint specification, we
rip out that codepath for %epic fact giving. For up-to-date ships, this
hasn't been in relevant use since the big update in #2817, about a year
and a half ago. That's been plenty of time for newer clients to send out
%epic facts to way-behind subscribers, too.

Messaging between ships should generally not be impacted by this. We
have fallback logic for the chat pokes, anyway.