discipline: remove own state from %egg-any imports
Agent exports obtained directly from gall will have the full state,
including that of wrapper agents. /lib/discipline is no exception, and
so needs to take care that when an agent gets an older egg imported into
it, that it remove the library state before passing the egg into the
inner agent.