Fixes issue where frontend state didn't reflect actual backend state after accepting a user join request
Removes "Send message" and "Copy user ID" actions from ProfileSheet, makes the header for the sheet pressable so that the user can go view the real UserProfileScreen that way.

Prevents user from assigning roles to an invited user (the state that the ship was stuck in on the frontend before the fix) while the user is still in the "invited" state.

Updates the GroupJoinRequest to actually go to the profile when the user presses "View profile" button, and removes the "Send message" and "Copy ID" action buttons (which are accessible in the actual UserProfileScreen).