channels: retrieve tombstones from hosts
On-upgrade, in addition to bunting all tombstones, we poke
channels-server instances with the request to send us their tombstones.
Very similar to the behavior added for re-syncing sequence numbers.

We update channels agent tests for the new types, update the sequence nr
retrieval test for the new behavior (where we ignore tombstones), and
add a new test for this new tombstone retrieval flow.