channels: modern conversion lib and new endpoints
The main focus of this commit is to add new version 4 endpoints to all
of our scries/subs. To do so I added a new conversion library similar
to what Mikolaj has done for the groups refactor.

Additionally this changes how we reference versioned types to use something
like v9:c which is much cleaner than v9:old:c. This also tries to use a
referenced version everywhere we would have just used :c so that when
version updates come later, we don't need to change anything. This was
not comprehensive however so there are still some instances that should
be updated.