channels: prevent duplicate post sending
The client does not always have enough information to know whether its
message send made it into the backend, so it retries rather
aggressively. This may result in messages being sent multiple times.

Here, we make the backend track the sent-at timestamps for messages. The
client guarantees that each has a unique sent-at timestamp, so we can
use those to detect and prevent duplicate messages from being sent.

Even though it's ephemeral state, we must unfortunately add it into the
$state type proper. The alternative is threading is all the way through
into and out of the helper core, which would be less than pretty.

Includes relevant tests.