channels: sane in presence of posts with seq=0
Due to the way tombstone migrations and resyncing works, we might be in
a state where tombstones all have seq=0. That was tripping up the /range
query logic, making it think it was already past the range when
traversing the post list.

Here, we special-case seq=0 and only "early-out" if a post with a real
sequence number is out of range.