groups: resub instead of crash when inconsistent
Logic for processing group updates sometimes checks whether the update
makes sense for the state that we have. Previously, if an inconsistency
was detected, it would crash.

Crashing results in a %kick, which we handle by resubscribing. But we
resubscribe at the same timestamp, causing us to receive the same
update, bringing us back where we started.

Here, we instead do a controlled unsubscribe and resubscribe in such a
way that we fetch the full log for the group. Since this _should_ always
start with a %create update, we will discard our (likely inconsistent)
group state and rebuild from scratch. That should bring us back in line
with the group host's state, resolving inconsistencies.